failed quote analytics
