+++ src/analytics/event.ts
@@ -586,9 +586,9 @@ export type EventProperties = {
     error_code: number | undefined;
     reason: string;
     inputAsset: { symbol: string; address: string; chainId: ChainId };
-    inputAmount: string | number;
+    inputAmount: string | number | undefined;
     outputAsset: { symbol: string; address: string; chainId: ChainId };
-    outputAmount: string | number;
+    outputAmount: string | number | undefined;
   };
 
   [event.browserTrendingDappClicked]: {
